Crate yy_engine

A library for solving Yin-Yang puzzles

Rules

Shade the grid 2 colors such that all cells of each cell are connected orthoganally and no 2 by 2 region is completely shaded either color.

Structs

YinYang

A representation of a yin-yang puzzle.

Enums

Deduction

Tracks if a strategy is sucessful.

YinYangError

Errors that can be generated when working with Yin-Yang puzzles.